refactor getAOs, remove componentDidUpdate
